ActorHelper moved to utils package. CorsAllowed setting reworked.
