+++ src/main/resources/reference.conf
@@ -24,10 +24,9 @@ scorex {
     # Hash of API key string
     #apiKeyHash = ""
 
-    # Enable/disable CORS support
-    corsAllowed = false
-
-    # Allowed origin line for cors header value
+    # Enable/disable CORS support.
+    # This is an optional param. It would allow cors in case if this setting is set.
+    # If this setting will be omitted cors will be prohibited.
     corsAllowedOrigin = "*"
 
     # request processing timeout
