+++ src/main/scala/scorex/core/utils/ActorHelper.scala
@@ -1,12 +1,15 @@
-package scorex.core.api.http
+package scorex.core.utils
 
-import akka.pattern.ask
 import akka.actor.ActorRef
+import akka.pattern.ask
 import akka.util.Timeout
 
 import scala.concurrent.Future
 import scala.reflect.ClassTag
 
+/**
+  * Helper that encapsulates ask patter for actors and returns Future[_]
+  */
 trait ActorHelper {
 
   def askActor[A: ClassTag](actorRef: ActorRef, question: Any)
